+++ vck/src/commonMain/kotlin/at/asitplus/wallet/lib/jws/ReconstructedSdJwtClaims.kt
@@ -1,0 +1,10 @@
+package at.asitplus.wallet.lib.jws
+
+import at.asitplus.wallet.lib.data.SelectiveDisclosureItem
+
+/**
+ * Contains all claims that have been successfully reconstructed from an [SdJwtSigned]
+ */
+data class ReconstructedSdJwtClaims(
+    val claims: Collection<SelectiveDisclosureItem>
+)