+++ vck/src/commonTest/kotlin/at/asitplus/wallet/lib/agent/AgentSdJwtTest.kt
@@ -18,7 +18,8 @@ import at.asitplus.wallet.lib.jws.SdJwtSigned
 import com.benasher44.uuid.uuid4
 import io.github.aakira.napier.Napier
 import io.kotest.core.spec.style.FreeSpec
-import io.kotest.inspectors.forAll
+import io.kotest.matchers.collections.shouldBeSingleton
+import io.kotest.matchers.collections.shouldHaveSingleElement
 import io.kotest.matchers.collections.shouldHaveSize
 import io.kotest.matchers.shouldBe
 import io.kotest.matchers.types.shouldBeInstanceOf
@@ -65,10 +66,12 @@ class AgentSdJwtTest : FreeSpec({
 
         val verified = verifier.verifyPresentation(vp.sdJwt, challenge)
             .shouldBeInstanceOf<Verifier.VerifyPresentationResult.SuccessSdJwt>()
-        verified.validatedItems shouldHaveSize 2
+        verified.reconstructed.claims shouldHaveSize 2
 
-        verified.validatedItems.first { it.claimName == CLAIM_GIVEN_NAME }.claimValue.content shouldBe "Susanne"
-        verified.validatedItems.first { it.claimName == CLAIM_DATE_OF_BIRTH }.claimValue.content shouldBe "1990-01-01"
+        verified.reconstructed.claims.first { it.claimName == CLAIM_GIVEN_NAME }
+            .claimValue.content shouldBe "Susanne"
+        verified.reconstructed.claims.first { it.claimName == CLAIM_DATE_OF_BIRTH }
+            .claimValue.content shouldBe "1990-01-01"
         verified.isRevoked shouldBe false
     }
 
@@ -84,8 +87,8 @@ class AgentSdJwtTest : FreeSpec({
         ).sdJwt
         val verified = verifier.verifyPresentation(sdJwt, challenge)
             .shouldBeInstanceOf<Verifier.VerifyPresentationResult.SuccessSdJwt>()
-        verified.validatedItems shouldHaveSize 1
-        verified.validatedItems.forAll { it.claimName shouldBe CLAIM_GIVEN_NAME }
+        verified.reconstructed.claims.shouldBeSingleton()
+        verified.reconstructed.claims.shouldHaveSingleElement { it.claimName == CLAIM_GIVEN_NAME }
         verified.isRevoked shouldBe false
     }